# comtea-front/src/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/communaute":{"get":{"tags":["communaute-api"],"operationId":"recupererCommunautes","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Communaute"}}}}}}},"post":{"tags":["communaute-api"],"operationId":"creerOuMettreAJourCommunaute","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Communaute"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/collaborateur":{"get":{"tags":["collaborateur-api"],"operationId":"recupererCollaborateurs","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Collaborateur"}}}}}}},"post":{"tags":["collaborateur-api"],"operationId":"creerCollaborateur","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Collaborateur"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/role-tournant/{role}/historique":{"get":{"tags":["role-tournant-api"],"operationId":"historique","parameters":[{"name":"role","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HistoriqueElection"}}}}}}}},"/api/role-tournant/{role}/election":{"get":{"tags":["role-tournant-api"],"operationId":"election","parameters":[{"name":"role","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Collaborateur"}}}}}}},"/api/communaute/{identifiant}":{"delete":{"tags":["communaute-api"],"operationId":"supprimerCommunaute","parameters":[{"name":"identifiant","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/collaborateur/{identifiant}":{"delete":{"tags":["collaborateur-api"],"operationId":"supprimerCollaborateur","parameters":[{"name":"identifiant","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"Communaute":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"identifiant":{"type":"string"},"nom":{"type":"string"}}},"Collaborateur":{"required":["identifiant"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"identifiant":{"type":"string"}}},"HistoriqueElection":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"role":{"type":"string"},"vainqueur":{"type":"string"},"date":{"type":"string","format":"date"}}}}}}